Revert "Merge remote-tracking branch 'upstream/develop' into 13p6"
This reverts commit 86aeb21f91f4e4f1bf0e7f46a34323381200ed01, reversing
changes made to 5726850cca3f33a47812a9d7b09da1f7fe9e7f55.

+++ htdocs/compta/bank/various_payment/card.php
@@ -60,9 +60,7 @@ if (!empty($conf->accounting->enabled) && !empty($conf->global->ACCOUNTANCY_COMB
 
 // Security check
 $socid = GETPOST("socid", "int");
-if ($user->socid) {
-	$socid = $user->socid;
-}
+if ($user->socid) $socid = $user->socid;
 $result = restrictedArea($user, 'banque', '', '', '');
 
 $object = new PaymentVarious($db);
@@ -77,37 +75,36 @@ $hookmanager->initHooks(array('variouscard', 'globalcard'));
 
 $parameters = array();
 $reshook = $hookmanager->executeHooks('doActions', $parameters, $object, $action); // Note that $action and $object may have been modified by some hooks
-if ($reshook < 0) {
-	setEventMessages($hookmanager->error, $hookmanager->errors, 'errors');
-}
+if ($reshook < 0) setEventMessages($hookmanager->error, $hookmanager->errors, 'errors');
 
-if (empty($reshook)) {
+if (empty($reshook))
+{
 	// Link to a project
-	if ($action == 'classin' && $user->rights->banque->modifier) {
+	if ($action == 'classin' && $user->rights->banque->modifier)
+	{
 		$object->fetch($id);
 		$object->setProject(GETPOST('projectid'));
 	}
 
-	if ($cancel) {
-		if ($action != 'addlink') {
+	if ($cancel)
+	{
+		if ($action != 'addlink')
+		{
 			$urltogo = $backtopage ? $backtopage : dol_buildpath('/compta/bank/various_payment/list.php', 1);
 			header("Location: ".$urltogo);
 			exit;
 		}
-		if ($id > 0 || !empty($ref)) {
-			$ret = $object->fetch($id, $ref);
-		}
+		if ($id > 0 || !empty($ref)) $ret = $object->fetch($id, $ref);
 		$action = '';
 	}
 
-	if ($action == 'add') {
+	if ($action == 'add')
+	{
 		$error = 0;
 
 		$datep = dol_mktime(12, 0, 0, GETPOST("datepmonth", 'int'), GETPOST("datepday", 'int'), GETPOST("datepyear", 'int'));
 		$datev = dol_mktime(12, 0, 0, GETPOST("datevmonth", 'int'), GETPOST("datevday", 'int'), GETPOST("datevyear", 'int'));
-		if (empty($datev)) {
-			$datev = $datep;
-		}
+		if (empty($datev)) $datev = $datep;
 
 		$object->ref = ''; // TODO
 		$object->accountid = GETPOST("accountid", 'int') > 0 ? GETPOST("accountid", "int") : 0;
@@ -129,42 +126,50 @@ if (empty($reshook)) {
 		$object->sens = GETPOSTINT('sens');
 		$object->fk_project = GETPOSTINT('fk_project');
 
-		if (empty($datep) || empty($datev)) {
+		if (empty($datep) || empty($datev))
+		{
 			$langs->load('errors');
 			setEventMessages($langs->trans("ErrorFieldRequired", $langs->transnoentitiesnoconv("Date")), null, 'errors');
 			$error++;
 		}
-		if (empty($object->amount)) {
+		if (empty($object->amount))
+		{
 			$langs->load('errors');
 			setEventMessages($langs->trans("ErrorFieldRequired", $langs->transnoentitiesnoconv("Amount")), null, 'errors');
 			$error++;
 		}
-		if (!empty($conf->banque->enabled) && !$object->accountid > 0) {
+		if (!empty($conf->banque->enabled) && !$object->accountid > 0)
+		{
 			$langs->load('errors');
 			setEventMessages($langs->trans("ErrorFieldRequired", $langs->transnoentitiesnoconv("BankAccount")), null, 'errors');
 			$error++;
 		}
-		if (empty($object->type_payment) || $object->type_payment < 0) {
+		if (empty($object->type_payment) || $object->type_payment < 0)
+		{
 			$langs->load('errors');
 			setEventMessages($langs->trans("ErrorFieldRequired", $langs->transnoentitiesnoconv("PaymentMode")), null, 'errors');
 			$error++;
 		}
-		if (!empty($conf->accounting->enabled) && !$object->accountancy_code) {
+		if (!empty($conf->accounting->enabled) && !$object->accountancy_code)
+		{
 			$langs->load('errors');
 			setEventMessages($langs->trans("ErrorFieldRequired", $langs->transnoentitiesnoconv("AccountAccounting")), null, 'errors');
 			$error++;
 		}
-		if ($object->sens < 0) {
+		if ($object->sens < 0)
+		{
 			$langs->load('errors');
 			setEventMessages($langs->trans("ErrorFieldRequired", $langs->transnoentitiesnoconv("Sens")), null, 'errors');
 			$error++;
 		}
 
-		if (!$error) {
+		if (!$error)
+		{
 			$db->begin();
 
 			$ret = $object->create($user);
-			if ($ret > 0) {
+			if ($ret > 0)
+			{
 				$db->commit();
 				$urltogo = ($backtopage ? $backtopage : DOL_URL_ROOT.'/compta/bank/various_payment/list.php');
 				header("Location: ".$urltogo);
@@ -179,23 +184,26 @@ if (empty($reshook)) {
 		$action = 'create';
 	}
 
-	if ($action == 'delete') {
+	if ($action == 'delete')
+	{
 		$result = $object->fetch($id);
 
-		if ($object->rappro == 0) {
+		if ($object->rappro == 0)
+		{
 			$db->begin();
 
 			$ret = $object->delete($user);
-			if ($ret > 0) {
-				if ($object->fk_bank) {
+			if ($ret > 0)
+			{
+				if ($object->fk_bank)
+				{
 					$accountline = new AccountLine($db);
 					$result = $accountline->fetch($object->fk_bank);
-					if ($result > 0) {
-						$result = $accountline->delete($user); // $result may be 0 if not found (when bank entry was deleted manually and fk_bank point to nothing)
-					}
+					if ($result > 0) $result = $accountline->delete($user); // $result may be 0 if not found (when bank entry was deleted manually and fk_bank point to nothing)
 				}
 
-				if ($result >= 0) {
+				if ($result >= 0)
+				{
 					$db->commit();
 					header("Location: ".DOL_URL_ROOT.'/compta/bank/various_payment/list.php');
 					exit;
@@ -231,18 +239,18 @@ if (empty($reshook)) {
 }
 
 // Action clone object
-if ($action == 'confirm_clone' && $confirm != 'yes') {
-	$action = '';
-}
+if ($action == 'confirm_clone' && $confirm != 'yes') { $action = ''; }
 
-if ($action == 'confirm_clone' && $confirm == 'yes' && ($user->rights->banque->modifier)) {
+if ($action == 'confirm_clone' && $confirm == 'yes' && ($user->rights->banque->modifier))
+{
 	$db->begin();
 
 	$originalId = $id;
 
 	$object->fetch($id);
 
-	if ($object->id > 0) {
+	if ($object->id > 0)
+	{
 		$object->id = $object->ref = null;
 
 		if (GETPOST('clone_label', 'alphanohtml')) {
@@ -253,30 +261,30 @@ if ($action == 'confirm_clone' && $confirm == 'yes' && ($user->rights->banque->m
 
 		$newdatepayment = dol_mktime(0, 0, 0, GETPOST('clone_date_paymentmonth', 'int'), GETPOST('clone_date_paymentday', 'int'), GETPOST('clone_date_paymentyear', 'int'));
 		$newdatevalue = dol_mktime(0, 0, 0, GETPOST('clone_date_valuemonth', 'int'), GETPOST('clone_date_valueday', 'int'), GETPOST('clone_date_valueyear', 'int'));
-		if ($newdatepayment) {
-			$object->datep = $newdatepayment;
-		}
+		if ($newdatepayment) $object->datep = $newdatepayment;
 		if (!empty($newdatevalue)) {
 			$object->datev = $newdatevalue;
 		} else {
 			$object->datev = $newdatepayment;
 		}
 
-		if (GETPOSTISSET("clone_sens")) {
-			$object->sens = GETPOST("clone_sens", 'int');
-		} else {
-			$object->sens = $object->sens;
-		}
+        if (GETPOSTISSET("clone_sens")) {
+            $object->sens = GETPOST("clone_sens", 'int');
+        } else {
+            $object->sens = $object->sens;
+        }
 
-		if (GETPOST("clone_amount", "alpha")) {
-			$object->amount = price2num(GETPOST("clone_amount", "alpha"));
-		} else {
-			$object->amount = price2num($object->amount);
-		}
+        if (GETPOST("clone_amount", "alpha")) {
+            $object->amount = price2num(GETPOST("clone_amount", "alpha"));
+        } else {
+            $object->amount = price2num($object->amount);
+        }
 
-		if ($object->check()) {
+		if ($object->check())
+		{
 			$id = $object->create($user);
-			if ($id > 0) {
+			if ($id > 0)
+			{
 				$db->commit();
 				$db->close();
 
@@ -308,17 +316,15 @@ if ($action == 'confirm_clone' && $confirm == 'yes' && ($user->rights->banque->m
 llxHeader("", $langs->trans("VariousPayment"));
 
 $form = new Form($db);
-if (!empty($conf->accounting->enabled)) {
-	$formaccounting = new FormAccounting($db);
-}
-if (!empty($conf->projet->enabled)) {
-	$formproject = new FormProjets($db);
-}
+if (!empty($conf->accounting->enabled)) $formaccounting = new FormAccounting($db);
+if (!empty($conf->projet->enabled)) $formproject = new FormProjets($db);
 
-if ($id) {
+if ($id)
+{
 	$object = new PaymentVarious($db);
 	$result = $object->fetch($id);
-	if ($result <= 0) {
+	if ($result <= 0)
+	{
 		dol_print_error($db);
 		exit;
 	}
@@ -339,7 +345,8 @@ foreach ($bankcateg->fetchAll() as $bankcategory) {
 /* Create mode                                                                */
 /*                                                                            */
 /* ************************************************************************** */
-if ($action == 'create') {
+if ($action == 'create')
+{
 	// Update fields properties in realtime
 	if (!empty($conf->use_javascript_ajax))
 	{
@@ -416,7 +423,8 @@ if ($action == 'create') {
 	print '</td></tr>';
 
 	// Bank
-	if (!empty($conf->banque->enabled)) {
+	if (!empty($conf->banque->enabled))
+	{
 		print '<tr><td>';
 		print $form->editfieldkey('BankAccount', 'selectaccountid', '', $object, 0, 'string', '', 1).'</td><td>';
 		$form->select_comptes($accountid, "accountid", 0, '', 2); // Affiche liste des comptes courant
@@ -485,7 +493,8 @@ if ($action == 'create') {
 	print '</td></tr>';
 
 	// Project
-	if (!empty($conf->projet->enabled)) {
+	if (!empty($conf->projet->enabled))
+	{
 		$formproject = new FormProjets($db);
 
 		// Associated project
@@ -530,24 +539,26 @@ if ($action == 'create') {
 /*                                                                            */
 /* ************************************************************************** */
 
-if ($id) {
+if ($id)
+{
 	$alreadyaccounted = $object->getVentilExportCompta();
 
 	$head = various_payment_prepare_head($object);
 
 	// Clone confirmation
-	if ($action === 'clone') {
-		$set_value_help = $form->textwithpicto('', $langs->trans($langs->trans("AccountingDirectionHelp")));
-		$sensarray = array('0' => $langs->trans("Debit"), '1' => $langs->trans("Credit"));
+	if ($action === 'clone')
+	{
+        $set_value_help = $form->textwithpicto('', $langs->trans($langs->trans("AccountingDirectionHelp")));
+        $sensarray = array('0' => $langs->trans("Debit"), '1' => $langs->trans("Credit"));
 
 		$formquestion = array(
 			array('type' => 'text', 'name' => 'clone_label', 'label' => $langs->trans("Label"), 'value' => $langs->trans("CopyOf").' '.$object->label),
-			array('type' => 'date', 'tdclass'=>'fieldrequired', 'name' => 'clone_date_payment', 'label' => $langs->trans("DatePayment"), 'value' => -1),
-			array('type' => 'date', 'name' => 'clone_date_value', 'label' => $langs->trans("DateValue"), 'value' => -1),
-			array('type' => 'other', 'tdclass'=>'fieldrequired', 'name' => 'clone_accountid', 'label' => $langs->trans("BankAccount"), 'value' => $form->select_comptes($object->fk_account, "accountid", 0, '', 1, '', 0, 'minwidth200', 1)),
-			array('type' => 'text', 'name' => 'clone_amount', 'label' => $langs->trans("Amount"), 'value' => price($object->amount)),
-			array('type' => 'select', 'name' => 'clone_sens', 'label' => $langs->trans("Sens") . ' ' . $set_value_help, 'values' => $sensarray, 'default' => $object->sens),
-		);
+            array('type' => 'date', 'tdclass'=>'fieldrequired', 'name' => 'clone_date_payment', 'label' => $langs->trans("DatePayment"), 'value' => -1),
+		    array('type' => 'date', 'name' => 'clone_date_value', 'label' => $langs->trans("DateValue"), 'value' => -1),
+		    array('type' => 'other', 'tdclass'=>'fieldrequired', 'name' => 'clone_accountid', 'label' => $langs->trans("BankAccount"), 'value' => $form->select_comptes($object->fk_account, "accountid", 0, '', 1, '', 0, 'minwidth200', 1)),
+		    array('type' => 'text', 'name' => 'clone_amount', 'label' => $langs->trans("Amount"), 'value' => price($object->amount)),
+            array('type' => 'select', 'name' => 'clone_sens', 'label' => $langs->trans("Sens") . ' ' . $set_value_help, 'values' => $sensarray, 'default' => $object->sens),
+        );
 
 		print $form->formconfirm($_SERVER["PHP_SELF"].'?id='.$object->id, $langs->trans('ToClone'), $langs->trans('ConfirmCloneVariousPayment', $object->ref), 'confirm_clone', $formquestion, 'yes', 1, 350);
 	}
@@ -556,10 +567,12 @@ if ($id) {
 
 	$morehtmlref = '<div class="refidno">';
 	// Project
-	if (!empty($conf->projet->enabled)) {
+	if (!empty($conf->projet->enabled))
+	{
 		$langs->load("projects");
 		$morehtmlref .= $langs->trans('Project').' ';
-		if ($user->rights->banque->modifier) {
+		if ($user->rights->banque->modifier)
+		{
 			if ($action != 'classify') {
 				$morehtmlref .= '<a class="editfielda" href="'.$_SERVER['PHP_SELF'].'?action=classify&amp;id='.$object->id.'">'.img_edit($langs->transnoentitiesnoconv('SetProject')).'</a> : ';
 			}
@@ -609,11 +622,7 @@ if ($id) {
 	print '</td></tr>';
 
 	// Debit / Credit
-	if ($object->sens == '1') {
-		$sens = $langs->trans("Credit");
-	} else {
-		$sens = $langs->trans("Debit");
-	}
+	if ($object->sens == '1') $sens = $langs->trans("Credit"); else $sens = $langs->trans("Debit");
 	print '<tr><td>'.$langs->trans("Sens").'</td><td>'.$sens.'</td></tr>';
 
 	print '<tr><td>'.$langs->trans("Amount").'</td><td>'.price($object->amount, 0, $outputlangs, 1, -1, -1, $conf->currency).'</td></tr>';
@@ -622,7 +631,8 @@ if ($id) {
 	print '<tr><td class="nowrap">';
 	print $langs->trans("AccountAccounting");
 	print '</td><td>';
-	if (!empty($conf->accounting->enabled)) {
+	if (!empty($conf->accounting->enabled))
+	{
 		$accountingaccount = new AccountingAccount($db);
 		$accountingaccount->fetch('', $object->accountancy_code, 1);
 
@@ -639,8 +649,10 @@ if ($id) {
 	print $form->editfieldval('SubledgerAccount', 'subledger_account', $object->subledger_account, $object, (!$alreadyaccounted && $user->rights->banque->modifier), 'string', '', 0);
 	print '</td></tr>';
 
-	if (!empty($conf->banque->enabled)) {
-		if ($object->fk_account > 0) {
+	if (!empty($conf->banque->enabled))
+	{
+		if ($object->fk_account > 0)
+		{
 			$bankline = new AccountLine($db);
 			$bankline->fetch($object->fk_bank);
 
@@ -675,13 +687,16 @@ if ($id) {
 	// Add button modify
 
 	// Clone
-	if ($user->rights->banque->modifier) {
+	if ($user->rights->banque->modifier)
+	{
 		print '<div class="inline-block divButAction"><a class="butAction" href="'.dol_buildpath("/compta/bank/various_payment/card.php", 1).'?id='.$object->id.'&amp;action=clone">'.$langs->trans("ToClone")."</a></div>";
 	}
 
 	// Delete
-	if (empty($object->rappro)) {
-		if (!empty($user->rights->banque->modifier)) {
+	if (empty($object->rappro))
+	{
+		if (!empty($user->rights->banque->modifier))
+		{
 			if ($alreadyaccounted) {
 				print '<div class="inline-block divButAction"><a class="butActionRefused classfortooltip" href="#" title="'.$langs->trans("Accounted").'">'.$langs->trans("Delete").'</a></div>';
 			} else {
